# Day 3: Load Testing Framework & Performance Validation